requirements.txt
annotated-types==0.7.0
asgiref==3.8.1
django==5.0.6
django-ninja==1.1.0
pydantic==2.7.4
pydantic-core==2.18.4
sqlparse==0.5.0
typing-extensions==4.12.2
